pyproject.toml
[project]
name = "token-vae"
version = "0.1.0"
description = "Token VAE for walkable manifold embedding space"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "sentencepiece>=0.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "datasets>=4.5.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "ruff>=0.2.0"]
viz = ["matplotlib>=3.8.0", "seaborn>=0.13.0", "umap-learn>=0.5.5", "scikit-learn>=1.4.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/token_vae"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
